+++ di/src/main/java/woowacourse/shopping/di/annotation/DependencyInjectAnnotation.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.di
+package woowacourse.shopping.di.annotation
 
 @Target(AnnotationTarget.PROPERTY)
 annotation class FieldInject
