+++ di/src/test/java/woowacourse/shopping/di/FakeDependencyObjects.kt
@@ -1,5 +1,7 @@
 package woowacourse.shopping.di
 
+import woowacourse.shopping.di.annotation.FieldInject
+
 interface ToBeInjected
 
 class FirstDependency : ToBeInjected
